Derhabicht/web 25/mkdirs (#23)
* Refactored handling of files on the disk

* feat: Building CAP file plans

New code added to create a file plan IAW CAPR 10-2. The input is a YAML
definition of the file plan, the output is a CSV that can be used to
generate CAPR 10-2 file labels and a PDF file plan formatted IAW the
example in CAPR 10-2 (Fig. 1).

* Debugged install directories and new file utility

* fix: damx no longer pulls in dpdocs

Also upped the loglevel on the damx installer

* Final debugging for DAMX completed

* Added attribute to not make a label in fileplan

* Formatting PDF fileplan by hierarchy

Top-level entries are bolded and underlined, guide cards are bold,
hanging file folders are italicized, and regular file folders are in
normal font.

* fix: Grouping on fileplan

* Added a Personnel File directory creation utility

I would like to expand this out to a full personnel file sync utility,
so it's not ready to be merged into `main` yet.

* Adding some metadata in anticipation of syncing

* chore: Factor out files and latex internal libs

The code that used to live in internal/files and internal/latex now live
at github.com/ag7if/go-files and github.com/ag7if/go-latex.

* Starting to integrate file plan w/personnel files